Feat: Conversação com IA, Ajuste em verfificação e salvamento em nuvem

+++ FuriaTec_TelegramBot/Program.cs
@@ -11,8 +11,10 @@ using RestSharp;
 using Newtonsoft.Json;
 using System.Text.RegularExpressions;
 using EsportsModels;
+using LibreOpenAI.OpenAi.ChatAi;
+using System.Reflection;
 
-
+var treinadorEscolhido = "";
 var serviceProvider = new ServiceCollection()
             .AddSingleton<ChatService>()
             .AddSingleton<ImagemService>()
@@ -27,7 +29,7 @@ var cadastroService = serviceProvider.GetService<CadastroService>();
 
 
 using var cts = new CancellationTokenSource();
-var bot = new TelegramBotClient("7915637811:AAEZWBusA9YKlHpj1oaDYSWMcrn9p2IA-hw", cancellationToken: cts.Token);
+var bot = new TelegramBotClient("", cancellationToken: cts.Token);
 var me = await bot.GetMe();
 
 
@@ -41,7 +43,7 @@ cts.Cancel();
 
 async Task OnError(Exception exception, HandleErrorSource source)
 {
-    Console.WriteLine(exception); // just dump the exception to the console
+     Console.WriteLine(exception); // just dump the exception to the console
 }
 
 async Task OnMessage(Message msg, UpdateType type)
@@ -49,48 +51,52 @@ async Task OnMessage(Message msg, UpdateType type)
     var telegramId = msg.From.Id;
     var nome = msg.From.FirstName;
     var texto = msg.Text;
-
-    if (!usuariosDB.UsuarioExiste(telegramId))
+    var time = 0;
+    if (msg.Text != null)
     {
-        if(cadastroService.EstaAguardandoEmail(telegramId))
+        if (!usuariosDB.UsuarioExiste(telegramId))
         {
+            if (cadastroService.EstaAguardandoEmail(telegramId))
+            {
+                usuariosDB.AdicionarUsuario(telegramId, nome, texto);
+                cadastroService.RemoverAguardando(telegramId);
+                time = 1;
+                await bot.SendMessage(msg.Chat.Id, "✅ Obrigado! Agora você faz parte da Nação FURIA!");
+                
+            }
 
-            usuariosDB.AdicionarUsuario(telegramId, nome, texto);
-            cadastroService.RemoverAguardando(telegramId);
-            await bot.SendMessage(msg.Chat.Id, "✅ Obrigado! Agora você faz parte da Nação FURIA!. Agora é so dar um /start e ir mais fundo na nossa equipe 🔥");
-        
+            // 2. Se não existe no banco, pede o e-mail
+            if (!usuariosDB.UsuarioExiste(telegramId))
+            {
+                cadastroService.MarcarComoAguardando(telegramId);
+                await bot.SendMessage(msg.Chat.Id, "🔥 Que bom que você chegou! Temos muito mais esperando por você. Montamos tudo com muito carinho para te deixar por dentro de tudo da Nação FURIA. 😎 Mas antes de começarmos, para você ficar por dentro de todas as novidades e ser um fã real da FURIA, manda seu e-mail aí para a gente! 📧");
+                return;
+            }
+
         }
-        else
+
+        if (time == 1)// Primeira vez
         {
-            
-            cadastroService.MarcarComoAguardando(telegramId);
-            await bot.SendMessage(msg.Chat.Id, "🔥 Que bom que você chegou! Temos muito mais esperando por você. Montamos tudo com muito carinho para te deixar por dentro de tudo da Nação FURIA. 😎 Mas antes de começarmos, para você ficar por dentro de todas as novidades e ser um fã real da FURIA, manda seu e-mail aí para a gente! 📧\"");
+            // Toca o áudio de boas-vindas
+            var buffer = File.ReadAllBytes("C:/Users/eduar/OneDrive/Documentos/Projetos/FuriaTec_TelegramBot/FuriaTec_TelegramBot/Resources/audio-fallen-furia.mp3");
+            await using var ms = new MemoryStream(buffer);
+            var message = await bot.SendAudio(msg.Chat, InputFile.FromStream(ms, "Bem Vindo.mp3"));
+            await bot.SendMessage(msg.Chat, "🐾 Fala, fã da FURIA!\r\nA FURIA é mais que um time — é uma nação movida por garra, paixão e atitude! 💥\r\nCada vitória, cada desafio, tudo é vivido junto com você, que faz parte dessa família insana. 💜\r\nManda aqui seu nome e receba um presente feito com todo o respeito e amor que a FURIA tem pelos seus fãs. 🎁🔥\r\nBora mostrar que #SerFURIA é viver intensamente!");
 
-            return;
+            var nomeCertificado = msg.From.FirstName;
+
+            // Gera e envia imagem
+            var caminhoImagem = await imagemService.GerarImagem(nomeCertificado);
+            var bufferImage = File.ReadAllBytes(caminhoImagem);
+            await using var msImg = new MemoryStream(bufferImage);
+            await bot.SendPhoto(msg.Chat.Id, InputFile.FromStream(msImg, "Certificado.png"));
+            bot.SendMessage(msg.Chat.Id, "🐾 Vem conhecer as opçoes feitas pra você Fâ, é só dar um\n/menu 🚀 ");
+
         }
-
-    }
-
-    if (!usuariosDB.UsuarioExiste(telegramId)) // Primeira vez
-    {
-        // Toca o áudio de boas-vindas
-        var buffer = File.ReadAllBytes("C:/Users/eduar/OneDrive/Documentos/Projetos/FuriaTec_TelegramBot/FuriaTec_TelegramBot/Resources/audio-fallen-furia.mp3");
-        await using var ms = new MemoryStream(buffer);
-        var message = await bot.SendAudio(msg.Chat, InputFile.FromStream(ms, "Bem Vindo.mp3"));
-        await bot.SendMessage(msg.Chat, "🐾 Fala, fã da FURIA!\r\nA FURIA é mais que um time — é uma nação movida por garra, paixão e atitude! 💥\r\nCada vitória, cada desafio, tudo é vivido junto com você, que faz parte dessa família insana. 💜\r\nManda aqui seu nome e receba um presente feito com todo o respeito e amor que a FURIA tem pelos seus fãs. 🎁🔥\r\nBora mostrar que #SerFURIA é viver intensamente!"); 
-
-        var nomeCertificado = msg.Text;
-
-        // Gera e envia imagem
-        var caminhoImagem = await imagemService.GerarImagem(nomeCertificado);
-        var bufferImage = File.ReadAllBytes(caminhoImagem);
-        await using var msImg = new MemoryStream(bufferImage);
-        await bot.SendPhoto(msg.Chat.Id, InputFile.FromStream(msImg, "Certificado.png"));
-    }
-    else
-    {
-  
-        await bot.SendMessage(msg.Chat.Id, "🐾 Eai, precisando de ajuda ?, é só dar um\n/menu 🚀 ");
+        else if (time == 1)
+        {
+            bot.SendMessage(msg.Chat.Id, "🐾 Eai, precisando de ajuda ?, Caso queria conhcer as opçoes ou voltar, é só dar um\n/menu 🚀 ");
+        }
     }
 
     switch (msg.Text)
@@ -99,18 +105,36 @@ async Task OnMessage(Message msg, UpdateType type)
 
             await bot.SendMessage(msg.Chat, "Você está prestes a treinar com os melhores da FURIA! 💪 \n Escolha seu treinador e prepare-se para evoluir no CS2. E se tiver dúvidas, é só perguntar estamos aqui pra te ajudar a dar o próximo passo! 🚀");
             
-            await bot.SendMessage(msg.Chat, "Escolha seu treinador:", replyMarkup: new InlineKeyboardMarkup(new[] {
+            var escolha = await bot.SendMessage(msg.Chat, "Escolha seu treinador:", replyMarkup: new InlineKeyboardMarkup(new[] {
                 new [] {
                     InlineKeyboardButton.WithCallbackData("FalleN", "fallen"),
                     InlineKeyboardButton.WithCallbackData("Yuurih", "yuurih"),
-                    InlineKeyboardButton.WithCallbackData("KSCERATO", "KSCERATO"),
-                    InlineKeyboardButton.WithCallbackData("Molodoy", "Molodoy"),
                     InlineKeyboardButton.WithCallbackData("Sidde", "Sidde")
-
                     }
             }));
 
-        break;
+
+            bot.OnUpdate += async (update) =>
+            {
+                if (update.CallbackQuery != null)
+                {
+                    var chatId = update.CallbackQuery.Message.Chat.Id;
+                    
+                    treinadorEscolhido = update.CallbackQuery.Data;
+                    await bot.SendMessage(chatId, $"Agora que você escolheu o {update.CallbackQuery.Data}, vamos juntos rumo à evolução no CS2!\n🚀 Com a experiência de um dos maiores nomes do cenário, você está mais perto de alcançar o topo!\nO que você quer aprimorar? 💥.");
+                }
+            };
+
+            bot.OnMessage += async (message, type) =>
+            {
+                var msg = message.Text;
+
+                var resposta = await chatService.PergunteAoTreinador(treinadorEscolhido, message.Text);
+                await bot.SendMessage(message.From.Id, resposta);
+            };
+
+            break;
+
 
         case "/jogos":
 
@@ -153,18 +177,10 @@ async Task OnMessage(Message msg, UpdateType type)
     }
 
 }
-async Task OnUpdate(Update update, Message msg)
+async Task OnUpdate(Update update)
 {
-    if (update is { CallbackQuery: { } query }) // non-null CallbackQuery
-    {
-        var escolhido = update.CallbackQuery.Data;
-        var chatId = update.CallbackQuery.Message.Chat.Id;
-        await bot.SendMessage(chatId, $"Agora que você escolheu o {escolhido}, vamos juntos rumo à evolução no CS2! 🚀 Com a experiência de um dos maiores nomes do cenário, você está mais perto de alcançar o topo!  O que você quer aprimorar? 💥.");
+    return;
 
-        var pergunta = msg.Text;
-        var resposta = await chatService.PergunteAoTreinador(escolhido, pergunta);
-        await bot.SendMessage(chatId, resposta);
-    }
 }
